requirements.txt
Flask
Flask-CORS
opencv-python
mediapipe
tensorflow
numpy
Pillow
deep-translator
google-translator
